+++ Resources/Public/JavaScript/record-translation-context-menu.js
@@ -1,0 +1,7 @@
+import RecordTranslationTrigger from './record-translation-trigger.js';
+
+export default {
+    triggerRecordTranslation(table, uid) {
+        RecordTranslationTrigger.trigger(table, parseInt(uid, 10));
+    },
+};